consolidate definitions of sizeof_fmt
